requirements.txt
Flask==1.0.3
Flask-RESTful==0.3.7
httpie==1.0.3
Flask-SQLAlchemy==2.4.0
Flask-Migrate==2.5.2
psycopg2-binary==2.8.3
marshmallow==2.19.5
webargs==5.4.0
markdown==2.6.11